Modificando para utilizar ticks no escalonamento e ajeitando lógica dos escalonamentos

+++ source/saida.h
@@ -6,15 +6,15 @@
 
 void criar_arquivo_saida(char tipo);
 
-void tarefa_finalizada(char tipo, double tempo, char nome[100]);
+void tarefa_finalizada(char tipo, int tempo, char nome[100]);
 
-void tarefa_interrompida(char tipo, double tempo, char nome[100]);
+void tarefa_interrompida(char tipo, int tempo, char nome[100]);
 
-void deadline_atingida(char tipo, double tempo, char nome[100]);
+void deadline_atingida(char tipo, int tempo, char nome[100]);
 
-void tempo_acabou(char tipo, double tempo, char nome[100]);
+void tempo_acabou(char tipo, int tempo, char nome[100]);
 
-void idle(char tipo, double tempo);
+void idle(char tipo, int tempo);
 
 void valores_tarefa_finalizada(char tipo, Atividade *head);
